refactor: storeInfo 파일명 변경
