+++ frontend/src/components/seller/signup/index.ts
@@ -1,3 +1,3 @@
 export { default as ProgressBar } from "./ProgressBar";
-export * from "./storeInfo";
+export * from "./stInfo";
 export * from "./common";
