+++ frontend/src/components/seller/signup/stInfo/RegisterDesc.tsx
