+++ frontend/src/components/seller/signup/stInfo/RegisterName.tsx
