+++ frontend/src/components/seller/signup/stInfo/index.ts
